+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+- Add `ObsClientSocket.Connected` event.
+- Rename `ObsClientSocket.Closed` event to `Disconnected`.
+
 ## 0.1.2
 
 - Add `(ChannelReader<IEvent>)ObsClientSocket.Events`. It can be used when `new ObsClientSocket(useChannel: true)`.